+++ Core/Interfaces/IConfigService.cs
@@ -66,4 +66,3 @@ namespace ArdysaModsTools.Core.Interfaces
         void Save();
     }
 }
-